requirements.txt
opencv-python>=4.9.0,<5.0.0
pyautogui>=0.9.54
pyperclip>=1.8.2
Pillow>=9.5.0
numpy>=1.26.0
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.24.0,<0.28.0
httpx>=0.25.0,<0.28.0
